last commit.
Project is done but I will be incorporating more stuff like edit products and orders.

+++ pages/api/products/[id].js
@@ -4,8 +4,11 @@ import Product from "../../../models/Product";
 export default async function handler(req, res) {
   const {
     method,
+    cookies,
     query: { id },
   } = req;
+
+  const token = cookies.token;
 
   await dbConnect();
 
@@ -18,6 +21,9 @@ export default async function handler(req, res) {
     }
   }
   if (method === "PUT") {
+    if (!token || token !== process.env.TOKEN) {
+      return res.status(401).json("Not Authenticated!");
+    }
     try {
       const product = await Product.findByIdAndUpdate(id, req.body, {
         new: true,
@@ -28,6 +34,9 @@ export default async function handler(req, res) {
     }
   }
   if (method === "DELETE") {
+    if (!token || token !== process.env.TOKEN) {
+      return res.status(401).json("Not Authenticated!");
+    }
     try {
       await Product.findByIdAndDelete(id);
       res.status(200).json("The product has been deleted!");
